% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acorn.R
\name{get_acorn}
\alias{get_acorn}
\title{Downloads CSV files from ACORN}
\usage{
get_acorn(dest)
}
\arguments{
\item{dest}{Path of the folder where the CSVs are to be extracted to}
}
\value{
A new folder with the data as CSV files
}
\description{
This functions pulls and extracts all the MAX Temp CSV files from all ACORN Stations.
}
\examples{
\dontrun{
get_acorn("destination_folder")
}

}
