% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_acorn.R
\name{merge_acorn}
\alias{merge_acorn}
\title{Title}
\usage{
merge_acorn(dir)
}
\arguments{
\item{dir}{Path to the folder of interest}
}
\value{
Tibble
}
\description{
Title
}
\examples{
\dontrun{
merge_acorn("path/to")
}
}
